requirements.txt
torch
torchvision
matplotlib
opencv-python
numpy
attrdict
IPython
pandas
pyyaml